FIX: EMMC Busy now ignores CRC ACK/NAK sequences
This fixes the busy wire coming out of the front end.  Prior to this
fix, an ACK/NAK could prevent the IP from detecting a busy.  Now,
the ACK/NAK will be ignored and the busy noticed following.

The SDTXFRAME has also been upgraded to produce errors on a CRC NAK
message.  Likewise, SDWB and SDAXIL now look for these NAK messages.
The spec.pdf has been upgraded to discuss this new interface.

Note that the ACK/NAK sequences are not (yet) recognized by the front
end.  Once they are, the logic following should work.

Also: Refreshed the wb2axip IP, and updated the copyright dates to
reflect the 2024 work taking place.